REST : Split `RestUpgradeAction` into `RestUpgradeAction` and `RestUpgradeStatusAction` fixes #29062
